Patch from Florian

+++ bench/sequential/ammunition/ammunition_libc.c
@@ -114,7 +114,7 @@ int ammunition_sprintf_d( char *s, int number )
   /* How many decimal digits do we need? */
   char digits = 0;
   unsigned char writePos = 0;
-  long copyOfNumber = number;
+  long long copyOfNumber = number;
   _Pragma( "loopbound min 1 max 10" )
   do {
     digits++;
